moved click logic into onElementSelect (outside CountryListRow)
